+++ Source/DOM classes/Core DOM/Text.h
@@ -18,6 +18,6 @@
 
 - (id)initWithValue:(NSString*) v;
 
--(Text*) splitText:(unsigned long) offset;
+-(Text*) splitText:(NSUInteger) offset;
 
 @end
